Add files via upload
Overfitting is demonstrated when a machine learning model can not categorize data well either because there is too many details or noise within the data.  
Underfitting is demonstrated when there  is not enough data for a model to train well on and it can not find the trend of the data. This is seen when a linear model is being constructed with non-linear data. 
Good fit is the ideal model because predictions are made with little to no errors. The model's good performance is learned from the training data.